working on delete issue

+++ backend/mutations/ThreadMutations.js
@@ -36,16 +36,19 @@ const deleteThread = {
     type: ThreadType,
     args: {
         threadId:{type: GraphQLNonNull(GraphQLID), resolve:async()=>{
-            let thread = await Thread.find({_id: args.thread})
+            let thread = await Thread.find({_id: args.threadId})
             return thread;
         }}
     },
     async resolve(parentValue,args, {user}){
-        return Thread.deleteOne({_id: args.threadId}).then(result=>{
-            return{_id: args.threadId}
-        }).catch(err=>{
-            throw err
-        })
+        let getThread = await Thread.findById(args.threadId)
+        if(getThread.user == user.id){
+            return Thread.deleteOne({_id: args.threadId}).then(result=>{
+                return{_id: args.threadId}
+                }).catch(err=>{
+                    throw err
+            })
+        }
     }
 }
 const upvoteThread = {
